+++ src/image_manager/ImageManager.hpp
@@ -37,7 +37,9 @@ private:
     void processImage(const InputImage& image);
     void issueWarningIfIsCentralizedRepositoryAndIsNotRootUser() const;
     void issueErrorIfIsCentralizedRepositoryAndCentralizedRepositoryIsDisabled() const;
-    void printLog(const boost::format& message, common::LogLevel LogLevel) const;
+    void issueErrorIfPullingByDigest() const;
+    void printLog(const boost::format& message, common::LogLevel LogLevel,
+                  std::ostream& outStream=std::cout, std::ostream& errStream=std::cerr) const;
 
 private:
     std::shared_ptr<const common::Config> config;